echo-shiro/src/main/java/org/echo/shiro/web/session/mgt/ServletUtilWrapper.java
package org.echo.shiro.web.session.mgt;

import org.springframework.web.context.request.RequestContextHolder;
import org.springframework.web.context.request.ServletRequestAttributes;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

/**
 * @author Liguiqing
 * @since V1.0
 */

public class ServletUtilWrapper {

    private ServletUtilWrapper() {
        throw new AssertionError("No org.echo.shiro.web.session.mgt.ServletUtilWrapper instances for you!");
    }

    public static HttpServletRequest getRequest() {
        return ((ServletRequestAttributes) RequestContextHolder.getRequestAttributes()).getRequest();
    }

    public static HttpServletResponse getResponse() {
        return ((ServletRequestAttributes) RequestContextHolder.getRequestAttributes()).getResponse();
    }
}